# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления задачами с авторизацией и воркерами",
        "title": "Tasks API",
        "contact": {
            "name": "API Support",
            "email": "support@tasksapi.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/add": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Добавляет новую задачу в систему обработки",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Добавить задачу",
                "parameters": [
                    {
                        "description": "Данные задачи",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":\"access\",\"uuid\":\"string\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "{\"error\":\"should contain task\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "{\"error\":\"server is busy\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/delete": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет задачу по UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Удалить задачу",
                "parameters": [
                    {
                        "description": "UUID задачи",
                        "name": "uuid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.TaskID"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":\"access\",\"deleted\ttask\":\"string\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "{\"error\":\"string\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "{\"status\":\"Not\tfound\tcurrent\ttask\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/refresh": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет пару access и refresh токенов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Обновить токены",
                "parameters": [
                    {
                        "description": "Refresh токен",
                        "name": "refresh",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":\"string\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "{\"error\":\"string\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "{\"error\":\"invalid\tauthorization\tformat\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "{\"error\":\"string\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/register": {
            "get": {
                "description": "Создает нового пользователя и возвращает пару токенов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Зарегистрировать пользователя",
                "responses": {
                    "200": {
                        "description": "{\"status\":\"string\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "{\"error\":\"string\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/status": {
            "post": {
                "description": "Возвращает текущий статус задачи",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Получить статус задачи",
                "parameters": [
                    {
                        "description": "UUID задачи",
                        "name": "uuid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.TaskID"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":\"access\", \"task name\": \"string\", \"createdAt\": date, \"current status\": \"string\", \"working time\": \"diff time\" }",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "204": {
                        "description": "{\"status\":\"not found task\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "{\"error\":\"Bad request: should contain UUID\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "{\"error\":\"Not\tfound\tcurrent\ttask\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.RefreshRequest": {
            "type": "object",
            "required": [
                "refresh"
            ],
            "properties": {
                "refresh": {
                    "description": "Refresh токен для обновления пары\n@Example eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiMTIzIn0.ABC123...",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
        "handlers.Task": {
            "description": "Модель задачи для создания",
            "type": "object",
            "required": [
                "taskname"
            ],
            "properties": {
                "taskname": {
                    "description": "Название задачи\n@Example \"Провести код-ревью\"",
                    "type": "string",
                    "example": "Какая то длинная io bound"
                }
            }
        },
        "handlers.TaskID": {
            "description": "Идентификатор задачи в формате UUID",
            "type": "object",
            "required": [
                "uuid"
            ],
            "properties": {
                "uuid": {
                    "description": "Уникальный идентификатор задачи\n@Example 6ba7b810-9dad-11d1-80b4-00c04fd430c8",
                    "type": "string",
                    "example": "6ba7b810-9dad-11d1-80b4-00c04fd430c8"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}